requirements.txt
allure-pytest==2.13.4
allure-python-commons==2.13.4
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
msgpack==1.0.8
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.7.0
websockets==13.0.1
Faker
pyhumps
PyJWT